backend-app/requirements.txt
fastapi
uvicorn 
pyjwt
python-jose